% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{calc_last_user_sync}
\alias{calc_last_user_sync}
\title{Calculates the last sync date in the action log data for each interviewer}
\usage{
calc_last_user_sync(df)
}
\arguments{
\item{df}{Data frame consisting of one or more user action log.}
}
\value{
Data frame .
}
\description{
Returns a data frame. Columns contain Contains the user ID (\code{UserId}) and last sync date (\code{last_sync_date}).
Rows correspond to interviewers.
\code{NA} in \code{last_sync_date} mean that there was no sync action identified in the action log--that is between the \code{start} and \code{end} dates used to obtain the action logs.
}
