% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_all_types}
\alias{download_all_types}
\title{Download all export types for a given questionnaire.}
\usage{
download_all_types(
  qnr_id,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  path,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. Format: \code{qnr_id$version}}

\item{interview_status}{Status of interviews to export. Values: \code{c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")}}

\item{from}{Start date for interviews to export. date-time, UTC}

\item{to}{End date for interviews to export. date-time, UTC}

\item{access_token}{Access token to external storage. Relevant only if external cloud storage is destination for the export file.}

\item{refresh_token}{Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.}

\item{storage_type}{External storage type, if relevant. Values: \code{c("Dropbox", "OneDrive", "GoogleDrive")}}

\item{translation_id}{Translation ID for variable and value labels to include in export files.}

\item{include_meta}{Logical. If \code{TRUE}, include questionnaire metadata in export file.}

\item{path}{File path where export file should be downloaded}

\item{workspace}{Character. Name of the workspace whose data to download.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Downloads tab-delimited, Stata, SPSS, binary, DDI, and Paradata for a given questionnaire version.
}
