% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download data for a single export job.}
\usage{
download_data(
  qnr_id,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  path,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{qnr_id}{Questionnaire ID. Format: \code{qnr_id$version}}

\item{export_type}{Type of data to export. Values: \code{c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")}}

\item{interview_status}{Status of interviews to export. Values: \code{c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")}}

\item{from}{Start date for interviews to export. date-time, UTC}

\item{to}{End date for interviews to export. date-time, UTC}

\item{access_token}{Access token to external storage. Relevant only if external cloud storage is destination for the export file.}

\item{refresh_token}{Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.}

\item{storage_type}{External storage type, if relevant. Values: \code{c("Dropbox", "OneDrive", "GoogleDrive")}}

\item{translation_id}{Translation ID for variable and value labels to include in export files.}

\item{include_meta}{Logical. If \code{TRUE}, include questionnaire metadata in export file.}

\item{path}{File path where export file should be downloaded}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose data to download.}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Executes full download process for a single export job.
}
\details{
\enumerate{
\item \strong{Starts export job.} Specifies the parameters of the download job through the \code{qnr_id}, \code{export_type}, \code{interview_status},  \code{from}, \code{to}, \code{access_token}, \code{refresh_token}, \code{storage_type}, \code{translation_id}, and \code{include_meta} parameters
\item \strong{Gets regular updates on the progress of the export job.} Messages about progress.
\item \strong{Downloads file to desired destination.} Use the \code{path} parameter to specify where the export file should be saved.
}
}
