% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_matching}
\alias{download_matching}
\title{Downloads data for questionnaires that match}
\usage{
download_matching(
  matches,
  export_type,
  interview_status = "All",
  from = "",
  to = "",
  access_token = "",
  refresh_token = "",
  storage_type = "",
  translation_id = "",
  include_meta = TRUE,
  path,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{matches}{Atomic character vector. Character string to be detected in questionnaire title. May be a string, sub-string, or regex.}

\item{export_type}{Type of data to export. Values: \code{c("Tabular", "STATA", "SPSS", "Binary", "DDI", "Paradata")}}

\item{interview_status}{Status of interviews to export. Values: \code{c("All", "SupervisorAssigned", "InterviewerAssigned", "Completed", "RejectedBySupervisor", "ApprovedBySupervisor", "RejectedByHeadquarters", "ApprovedByHeadquarters")}}

\item{from}{Start date for interviews to export. date-time, UTC}

\item{to}{End date for interviews to export. date-time, UTC}

\item{access_token}{Access token to external storage. Relevant only if external cloud storage is destination for the export file.}

\item{refresh_token}{Refresh token to external storage.Relevant only if external cloud storage is destination for the export file.}

\item{storage_type}{External storage type, if relevant. Values: \code{c("Dropbox", "OneDrive", "GoogleDrive")}}

\item{translation_id}{Translation ID for variable and value labels to include in export files.}

\item{include_meta}{Logical. If \code{TRUE}, include questionnaire metadata in export file.}

\item{path}{File path where export file should be downloaded}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\description{
If the export file is ready, this function downloads it to the target directory specified in \code{path}. If the file is not ready, print a message describing the situation.
}
