% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{find_matching_qnrs}
\alias{find_matching_qnrs}
\title{Get list of questionnaires with a matching title.}
\usage{
find_matching_qnrs(
  matches,
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{matches}{Atomic character vector. Character string to be detected in questionnaire title. May be a string, sub-string, or regex.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame of matching questionnaires, if successful; warning message, if not successful.
}
\description{
Get a data frame of questionnaires that match the character vector provided.
}
\details{
If matching questionnaires are found, return a data frame with attributes. These attributes can be used, among other things, to find the necessary parameters for downloading any associated data.

If no matching questionnaires are found, return a warning message.

In both cases, print the outcome to the console for information.
}
