% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_all_user_logs}
\alias{get_all_user_logs}
\title{Obtains action logs for all interviewer users}
\usage{
get_all_user_logs(
  start,
  end,
  workspace = "primary",
  server = Sys.getenv("SUSO_SERVER"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{start}{Character vector. Date in YYY-MM-DD format}

\item{end}{Character vector. Date in YYY-MM-DD format}

\item{workspace}{Character. Workspace whose action logs to get.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame.
}
\description{
Performs workflow to obtain all action logs:
}
\details{
\enumerate{
\item Obtain the list of all interviewers
\item Fetch the action logs for each interviewers
}

Contains action log columns: time of event (\code{Time}) and description of event (\code{UserId}).
Also contains an identifier for the user: \code{UserID}, a GUID that can be used to add the user name.
Rows correspond to events in the action log.
\code{NA} values in \code{Time} and \code{Message} mean that there were no actions in action log between the \code{start} and \code{end} dates used to obtain the action logs.
}
