% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_last_sync_dates}
\alias{get_last_sync_dates}
\title{Compute the last sync date for each interviewer}
\usage{
get_last_sync_dates(
  start,
  end,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{start}{Character vector. Date in YYY-MM-DD format}

\item{end}{Character vector. Date in YYY-MM-DD format}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Workspace whose action logs to get.}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Data frame consisting of one or more user action log.
}
\description{
Conducts full workflow to get the last sync date for each interviewer:
}
\details{
\enumerate{
\item Obtain the list of all interviewers
\item Fetch the action logs for each interviewers
\item Determine the last sync date for each interviewer
}

Returns a data frame. Columns contain Contains the user ID (\code{UserId}) and last sync date (\code{last_sync_date}).
Rows correspond to interviewers.
\code{NA} in \code{last_sync_date} mean that there was no sync action identified in the action log--that is between the \code{start} and \code{end} dates used to obtain the action logs.
}
