% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip_to_dir}
\alias{unzip_to_dir}
\title{Unpack all zip files found in directory to same-named sub-directory}
\usage{
unzip_to_dir(dir)
}
\arguments{
\item{dir}{Character. Directory where zip files can be found}
}
\value{
Side-effect of creating a folder for each zip and
unpacking its contents there.
}
\description{
Applies \code{unpack_zip_to_dir()} to all zip files in directory
}
